pagination updates

+++ controllers/admin.js
@@ -1,4 +1,5 @@
-const { validationResult } = require('express-validator')
+const fileHelper = require('../util/file');
+const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
 // Renders the add product form page for admin users
@@ -163,6 +164,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -202,14 +204,21 @@ exports.getProducts = (req, res, next) => {
 // Ensures only the product owner can delete their own products by checking userId
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId)
+  .then(product => {
+    if (!product) {
+      return next(new Error('Product not found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    Product.deleteOne({_id: prodId, userId: req.user._id});
+  })
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
+  .catch(err => {
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 };
